:hammer: Move some expansion panels methods to the form utils to enable reuse
