+++ gradle.properties
@@ -1,4 +1,4 @@
-VERSION_NAME=1.7.12-SNAPSHOT
+VERSION_NAME=1.7.13-SNAPSHOT
 VERSION_CODE=1
 GROUP=org.smartregister
 POM_SETTING_DESCRIPTION=OpenSRP Client Native Form Json Wizard
